helm-swoop action
Implement an action to search below a subtree with `helm-swoop`.

Use case: Searching instead of reading for long entries
-------------------------------------------------------

Searching is a good alternative to reading when looking at notes,
particularly when one has long entries. In such cases,
when you want to search within an entry
rather than find entries that contain certain text,
`helm-swoop` is a better fit than *helm-org-rifle`
since one wants to see the surrounding content.

There is a question as to why you don't just go to a heading
and then carry out a helm-swoop, but *orgnav* is in
many ways defined as a complete navigation system.

Related use case: When headings are not descriptive
---------------------------------------------------

This feature can be used in a similar way as `helm-org-rifle`:
To find headings containing useful information
when the headings themselves are not sufficiently
descriptive by searching the text content of entries.
Carrying out additional searching represents a
precision-recall trade off.

It might make sense to support similarly functionality to
`helm-org-rifle` at some stage: a search that list those
headings that contain a string. I don't really feel that
I want an **action** for this, so much as I would want
to make `helm-org-rifle` style-searching (where content is
included as a search criteria for headings) a togglable
option, since the function of `orgnav`, `helm-org-rifle`'s
display are similar (namely find a heading and jump to it).

I would feel a little bad about swallowing `helm-org-rifle` in doing
so however :/.